+++ crud/package.json
@@ -1,7 +1,7 @@
 {
   "name": "crud-api",
   "version": "1.0.0",
-  "main": "index.js",
+  "main": "app.ts",
   "license": "MIT",
   "scripts": {
     "start:dev": "yarn nodemon app.ts"
